+++ single.php
@@ -4,7 +4,7 @@
 
 		<div id="inner-content" class="m-padding pull-l-1of12 pull-r-1of12 cf">
 
-			<main id="main" class="pull-l-1of12 pull-r-1of12 m-all cf" role="main" itemscope itemprop="mainContentOfPage" itemtype="http://schema.org/LocalBusiness">
+			<main id="main" class="pull-l-1of12 pull-r-1of12 m-all cf" role="main" itemscope itemType="http://schema.org/BlogPosting">
 
 			<?php if (have_posts()) : while (have_posts()) : the_post(); ?>
 
@@ -12,14 +12,11 @@
 					get_template_part( 'post-formats/format', get_post_format() );
 				
 					if ( types_render_field( "report-pdf" ) != null ) { ?>
-					<div class="pad-top cf">
-						<a id="cta-border-green" class=" d-1of2 green" href="<?php echo types_render_field( "report-pdf", array( 'raw' => true ) ); ?>">Download the report</a>
-					</div>
-						
-
+						<div class="pad-top cf">
+							<a id="cta-border-green" class=" d-1of2 green" href="<?php echo types_render_field( "report-pdf", array( 'raw' => true ) ); ?>">Download the report</a>
+						</div>
 					<?php } 
-
-
+					
 					get_template_part( 'post-formats/author-single' );
 				?>